+++ src/Modules/Machinery/V1/SimplePiston.cs
@@ -113,7 +113,7 @@ public class SimplePiston : UpdatableAndDeletable, IDrawable
 	public void AddToContainer(RoomCamera.SpriteLeaser sLeaser, RoomCamera rCam, FContainer? newContatiner)
 	{
 		foreach (var fs in sLeaser.sprites) fs.RemoveFromContainer();
-		try { (newContatiner ?? rCam.ReturnFContainer(_mc?.containerName ?? ContainerCodes.Items)).AddChild(sLeaser.sprites[0]); }
+		try { (newContatiner ?? rCam.ReturnFContainer((_mc?.containerName ?? ContainerCodes.Items).ToString())).AddChild(sLeaser.sprites[0]); }
 		catch { rCam.ReturnFContainer("Items").AddChild(sLeaser.sprites[0]); }
 
 	}
