+++ src/Prelude.cs
@@ -9,5 +9,10 @@ global using RWCustom;
 global using UnityEngine;
 global using static Pom.Pom;
 global using static RegionKit.Mod;
-global using static RegionKit.Utils;
+global using static RegionKit.Extras.CollectionTools;
+global using static RegionKit.Extras.ReflectionTools;
+global using static RegionKit.Extras.RngTools;
+global using static RegionKit.Extras.PrimitivesTools;
+global using static RegionKit.Extras.RainWorldTools;
+global using static RegionKit.Extras.LanguageHacks;
 global using static RWCustom.Custom;
